# api/docs/openapi-spec.yaml
openapi: 3.1.0
info:
  title: FloresYa API
  version: 1.0.0
  description: E-commerce API for flower delivery - Built with KISS principles
  contact:
    name: FloresYa Team
    email: support@floresya.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://floresya.vercel.app
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
          description: Always false for error responses
        error:
          type: string
          example: ValidationError
          description: Human-readable error class name
        message:
          type: string
          example: Validation failed. Please check your input.
          description: User-friendly error message (FloresYa extension)
        details:
          type: object
          items:
            type: string
          description: Additional context about the error (for operational errors)
          example:
            field: email
            value: invalid-email
        code:
          type: integer
          example: 1001
          description: Numeric error code from ERROR_CODES
        category:
          type: string
          enum:
            - validation
            - authentication
            - not_found
            - business
            - server
          example: validation
          description: Error category for grouping
        type:
          type: string
          format: uri
          example: https://api.floresya.com/errors/validation/validationfailed
          description: RFC 7807 - URI reference identifying the problem type
        title:
          type: string
          example: Validation Failed
          description: RFC 7807 - Short, human-readable summary
        status:
          type: integer
          example: 400
          description: RFC 7807 - HTTP status code
        detail:
          type: string
          example: Request validation failed
          description: RFC 7807 - Detailed explanation of the error (same as message)
        instance:
          type: string
          example: /errors/req_123456789
          description: RFC 7807 - URI reference identifying this specific occurrence
        timestamp:
          type: string
          format: date-time
          example: '2025-10-31T15:10:00.000Z'
          description: ISO 8601 timestamp when error occurred
        path:
          type: string
          example: /api/products
          description: Request path that caused the error
        requestId:
          type: string
          example: req_123456789
          description: Unique request identifier for tracking
        errors:
          type: object
          description: Field-specific validation errors (for validation errors only)
          example:
            email:
              - Must be a valid email address
            password:
              - Must be at least 8 characters
      description: RFC 7807 compliant error response with FloresYa extensions
      required:
        - success
        - error
        - code
        - message
    user:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          example: John Doe
        phone:
          type: string
          example: '+1234567890'
        role:
          type: string
          enum:
            - user
            - admin
          example: user
        email_verified:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    order:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        user_id:
          type: integer
          example: 5
        customer_email:
          type: string
          format: email
          example: maria@example.com
        customer_name:
          type: string
          example: María González
        customer_phone:
          type: string
          example: +58 412-1234567
        delivery_address:
          type: string
          example: Av. Principal, Caracas
        delivery_date:
          type: string
          format: date
          example: '2025-10-05'
        delivery_time_slot:
          type: string
          example: 10:00-12:00
        delivery_notes:
          type: string
          example: Llamar al llegar
        status:
          type: string
          enum:
            - pending
            - verified
            - preparing
            - shipped
            - delivered
            - cancelled
          example: pending
        total_amount_usd:
          type: number
          format: decimal
          example: 89.99
        total_amount_ves:
          type: number
          format: decimal
          example: 3599.6
        currency_rate:
          type: number
          format: decimal
          example: 40
        notes:
          type: string
          example: Ocasión especial
        admin_notes:
          type: string
          example: Cliente frecuente
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Resource ID
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items to return
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
...
paths:
...
  /api/orders/user/{userId}:
    get:
      tags:
        - Orders
      summary: Get orders by user
      description: Get all orders for a specific user (owner or admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: User ID
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: User orders retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Admin only - Returns paginated list of all users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: role
          in: query
          schema:
            type: string
            enum:
              - user
              - admin
          description: Filter by user role
        - name: email_verified
          in: query
          schema:
            type: boolean
          description: Filter by email verification status
        - name: search
          in: query
          schema:
            type: string
          description: Search in full_name and email (accent-insensitive, uses indexed normalized columns)
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Users
      summary: Create new user
      description: Create a new user account (public registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - full_name
              properties:
                email:
                  type: string
                  format: email
                full_name:
                  type: string
                  minLength: 2
                  maxLength: 255
                phone:
                  type: string
                  pattern: ^\+?[\d\s-()]+$
                role:
                  type: string
                  enum:
                    - user
                    - admin
                  default: user
                password_hash:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: User already exists
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Get user details by ID (owner or admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags:
        - Users
      summary: Update user
      description: Update user details (owner or admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  minLength: 2
                  maxLength: 255
                phone:
                  type: string
                  pattern: ^\+?[\d\s-()]+$
                role:
                  type: string
                  enum:
                    - user
                    - admin
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - Users
      summary: Delete user (soft delete)
      description: Admin only - Soft deletes a user (sets is_active to false)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/users/email/{email}:
    get:
      tags:
        - Users
      summary: Get user by email
      description: Admin only - Get user details by email address
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
          description: User email address
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/users/{id}/reactivate:
    patch:
      tags:
        - Users
      summary: Reactivate user
      description: Admin only - Reactivates a soft-deleted user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User reactivated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/users/{id}/verify-email:
    patch:
      tags:
        - Users
      summary: Verify user email
      description: Verify user email address (owner or admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/user'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...